+++ enterprise/cmd/frontend/auth/openidconnect/config.go
@@ -51,11 +51,11 @@ func init() {
 	conf.ContributeValidator(validateConfig)
 }
 
-func validateConfig(c conf.Unified) (problems []string) {
+func validateConfig(c conf.Unified) (problems conf.Problems) {
 	var loggedNeedsExternalURL bool
 	for _, p := range c.Critical.AuthProviders {
 		if p.Openidconnect != nil && c.Critical.ExternalURL == "" && !loggedNeedsExternalURL {
-			problems = append(problems, `openidconnect auth provider requires externalURL to be set to the external URL of your site (example: https://sourcegraph.example.com)`)
+			problems = append(problems, conf.NewCriticalProblem("openidconnect auth provider requires `externalURL` to be set to the external URL of your site (example: https://sourcegraph.example.com)"))
 			loggedNeedsExternalURL = true
 		}
 	}
@@ -64,7 +64,7 @@ func validateConfig(c conf.Unified) (problems []string) {
 	for i, p := range c.Critical.AuthProviders {
 		if p.Openidconnect != nil {
 			if j, ok := seen[*p.Openidconnect]; ok {
-				problems = append(problems, fmt.Sprintf("OpenID Connect auth provider at index %d is duplicate of index %d, ignoring", i, j))
+				problems = append(problems, conf.NewCriticalProblem(fmt.Sprintf("OpenID Connect auth provider at index %d is duplicate of index %d, ignoring", i, j)))
 			} else {
 				seen[*p.Openidconnect] = i
 			}